+++ src/main/java/org/unreal/modelrouter/config/core/ConfigurationValidator.java
@@ -231,7 +231,8 @@ public class ConfigurationValidator {
         return "random".equals(normalizedType)
                 || "round-robin".equals(normalizedType)
                 || "least-connections".equals(normalizedType)
-                || "ip-hash".equals(normalizedType);
+                || "ip-hash".equals(normalizedType)
+                || "latency".equals(normalizedType);
     }
 
     /**